verenc/cd00: Move basic scheme to package basic
