+++ verenc/cd00/basic/prover.go
@@ -1,4 +1,4 @@
-package cd00
+package basic
 
 import (
 	"bytes"
