+++ verenc/cd00/basic/verifier.go
@@ -1,4 +1,4 @@
-package cd00
+package basic
 
 import (
 	"bytes"
